% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appr.R
\name{appr_num_tested}
\alias{appr_num_tested}
\title{APPR Number Tested}
\usage{
appr_num_tested(sirs301, subject, end_year, location_name)
}
\arguments{
\item{sirs301}{a valid sirs301 object}

\item{subject}{c('ELA', 'Math')}

\item{end_year}{integer, year that you want to generate the table for}

\item{location_name}{name of the location to report on.}
}
\value{
data.frame
}
\description{
APPR Number Tested
}

