% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appr.R
\name{appr_performance_by_year_table}
\alias{appr_performance_by_year_table}
\title{A function to make the APPR performance vs 2+ year performance table.}
\usage{
appr_performance_by_year_table(sirs301, subject, end_year, location_name)
}
\arguments{
\item{sirs301}{a valid sirs301 object}

\item{subject}{c('ELA', 'Math', 'Sci')}

\item{end_year}{integer, year that you want to generate the table for}

\item{location_name}{name of the location to report on.}
}
\value{
list
}
\description{
A function to make the APPR performance vs 2+ year performance table.
}
