% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appr.R
\name{appr_pli}
\alias{appr_pli}
\title{APPR PLI table}
\usage{
appr_pli(sirs301, subject, end_year, location_name)
}
\arguments{
\item{sirs301}{a valid sirs301 object}

\item{subject}{c('ELA', 'Math', 'Sci')}

\item{end_year}{integer, year that you want to generate the table for}

\item{location_name}{name of the location to report on.}
}
\value{
prints to console only
}
\description{
APPR PLI table
}

