% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calculate_cohort}
\alias{calculate_cohort}
\title{caclculate cohort}
\usage{
calculate_cohort(grade, start_year, kind = "college_entry")
}
\arguments{
\item{grade}{student's grade level}

\item{start_year}{start of the year in question.  eg, 2015-2016 would be 2015.}

\item{kind}{c('college_entry', 'college_grad').  should the cohort be tagged by
the year the student *enters* college, or the year the student will
*graduate* college?}
}
\value{
numeric vector with student's cohort name
}
\description{
caclculate cohort
}

