% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_performance_bins.R
\name{cohort_performance_bins}
\alias{cohort_performance_bins}
\title{Cohort Performance Bins}
\usage{
cohort_performance_bins(sirs301_obj, studentids, subject)
}
\arguments{
\item{sirs301_obj}{a sirs301 object}

\item{studentids}{students you want to plot}

\item{subject}{c('ELA', 'Math', 'Sci')}
}
\value{
ggplot2 object
}
\description{
Cohort Performance Bins
}

