% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{limit_ela_math}
\alias{limit_ela_math}
\title{utility function to filter a sirs301 object}
\usage{
limit_ela_math(data, studentids, subjects, years)
}
\arguments{
\item{data}{sirs301$ela_math data object}

\item{studentids}{students you want to include in the performance stats}

\item{subjects}{c('ELA', 'Math')}

\item{year}{test year (ending academic year)}
}
\value{
tbl_df with matching data
}
\description{
utility function to filter a sirs301 object
}
