% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_performance_bins.R
\name{perf_level_order}
\alias{perf_level_order}
\title{perf_level_order}
\usage{
perf_level_order(x)
}
\arguments{
\item{x}{a performance level (1-4)}
}
\description{
helper function used by cohort_performance_bins to put performance levels in correct order
}

