% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirs301.R
\name{print.sirs301}
\alias{print.sirs301}
\title{print method for \code{sirs301} class}
\usage{
\method{print}{sirs301}(x, ...)
}
\arguments{
\item{x}{a \code{sirs301} object}

\item{...}{additional arguments}
}
\value{
some details about the object to the console.
}
\description{
prints to console
}
\details{
Prints a summary of the a \code{sirs301} object.
}
