% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirs301.R
\name{sirs301}
\alias{sirs301}
\title{Create a sirs301 object}
\usage{
sirs301(csvs, cohort_kind = "college_entry", verbose = FALSE, ...)
}
\arguments{
\item{csvs}{data frame of sirs 301 exports - eg output of
sirs301_import_csvs}

\item{verbose}{should sirs301 print status updates?  default is FALSE.}

\item{...}{additional arguments to pass to constructor functions}
}
\description{
\code{sirs301} creates a sirs301 object, enabling analysis and reporting.
}

