% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirs301_growth.R
\name{sirs301_ela_math_growth}
\alias{sirs301_ela_math_growth}
\title{Generate student-level growth scores from the SIRS301 tested report}
\usage{
sirs301_ela_math_growth(sirs301_ela_math, headline = TRUE)
}
\arguments{
\item{sirs301_ela_math}{dataframe of ela and math scores.  eg $ela_math slot on sirs301 object.}

\item{headline}{logical, return the headline stats.  If FALSE, will return less frequently used
columns like start_lep_duration and end_lep_duration.  default is TRUE}
}
\value{
dataframe, class sirs301_ela_math_growth
}
\description{
Generate student-level growth scores from the SIRS301 tested report
}
