% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirs301_growth.R
\name{sirs301_empty_scaffold}
\alias{sirs301_empty_scaffold}
\title{Crate empty scaffold for}
\usage{
sirs301_empty_scaffold(sirs301_ela_math)
}
\arguments{
\item{sirs301_ela_math}{dataframe of ela and math scores.  eg $ela_math slot on sirs301 object.}
}
\value{
dataframe with all possible growth windows per student/subject
}
\description{
helper function for `sirs301_ela_math_growth`
}

