% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirs301_growth.R
\name{sirs301_fill_scaffold}
\alias{sirs301_fill_scaffold}
\title{Fill in the growth scaffold with data}
\usage{
sirs301_fill_scaffold(sirs301_ela_math, scaffold, headline = TRUE)
}
\arguments{
\item{sirs301_ela_math}{dataframe of ela and math scores.  eg $ela_math slot on sirs301 object.}

\item{scaffold}{output of `sirs301_empty_scaffold`}

\item{headline}{logical, return the headline stats.  If FALSE, will return less frequently used
columns like start_lep_duration and end_lep_duration.  default is TRUE}
}
\value{
dataframe with matching student data for all possible combinations.
}
\description{
helper function for `sirs301_ela_math_growth`
}

