% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirs301_growth.R
\name{sirs301_growth_calcs}
\alias{sirs301_growth_calcs}
\title{run growth calculations given a completed scaffold}
\usage{
sirs301_growth_calcs(complete_scaffold)
}
\arguments{
\item{complete_scaffold}{dataframe, output of sirs301_fill_scaffold}
}
\value{
dataframe with growth calculations
}
\description{
run growth calculations given a completed scaffold
}
