% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirs301_import.R
\name{sirs301_import}
\alias{sirs301_import}
\title{Title}
\usage{
sirs301_import(path, verbose = TRUE)
}
\arguments{
\item{path}{directory of exports from L2RPT.  To get the files, run the SIRS-301 report, click
the number in the 'Total Tested' grid to get the student drilldown, and then use the
'View in Excel Options' > 'View in CSV format' option to write the data to a file.  Do not choose
'View in Excel 2007 data' - the files generated by the IBM Cognos system are subtly different than files
produced by MS Excel and can't be read by the `readxl` package.}

\item{verbose}{if TRUE, will generate `message()` updates.}
}
\value{
tbl_df data frame of student-level records
}
\description{
Title
}

