% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topline_performance_chart.R
\name{topline_performance_chart}
\alias{topline_performance_chart}
\title{Topline Performance Chart}
\usage{
topline_performance_chart(sirs301_obj, studentids, subject, years,
  by = c("test_subject"))
}
\arguments{
\item{sirs301_obj}{a sirs301 object}

\item{studentids}{students you want to plot}

\item{subject}{c('ELA', 'Math')}

\item{years}{}
}
\value{
ggplot2 object
}
\description{
Topline Performance Chart
}
