% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topline_stats.R
\name{topline_performance_stats}
\alias{topline_performance_stats}
\title{Topline Performance Stats}
\usage{
topline_performance_stats(sirs301_obj, studentids, subjects = c("ELA",
  "Math"), years = c(2017), by = c("district_name", "test_subject"))
}
\arguments{
\item{sirs301_obj}{a sirs301 object}

\item{studentids}{students you want to include in the performance stats}

\item{subjects}{c('ELA', 'Math')}

\item{year}{test year (ending academic year)}
}
\value{
ggplot2 object
}
\description{
Topline Performance Stats
}
